requirements.txt
# Data & ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.0.0

# Experiment tracking
mlflow>=3.0.0
setuptools

# API
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0

# Dashboard
streamlit>=1.36.0
plotly>=5.22.0

# Database
# sqlite3 is built into Python - no install needed

# LLM
groq>=0.9.0

# Monitoring
scipy>=1.14.0

# Testing
pytest>=8.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
